feat: 暂存

+++ actions/menu/index.ts
@@ -1,8 +1,8 @@
-'use server';
-import { auth } from '@/auth';
-import { DocumentVO } from '@/shared';
-import { createClient } from '@/supabase/server';
-import { cookies } from 'next/headers';
+"use server";
+import { auth } from "@/auth";
+import { DocumentVO } from "@/shared";
+import { createClient } from "@/supabase/server";
+import { cookies } from "next/headers";
 
 export async function getMenus() {
   const session = await auth();
@@ -10,11 +10,10 @@ export async function getMenus() {
   const cookieStore = await cookies();
   const supabase = createClient(cookieStore);
   const { data } = await supabase
-    .from('document_v2')
-    .select('*')
-    .order('create_at', { ascending: true })
-    .eq('creator', session.user.name)
-    .eq('creator_email', session.user.email);
+    .from("document_v2")
+    .select("*")
+    .order("create_at", { ascending: true })
+    .eq("uid", session.user.id);
 
   return (data || []) as DocumentVO[];
 }
@@ -25,11 +24,11 @@ export async function createDoc(id: string, parent_id?: string) {
   const cookieStore = await cookies();
   const supabase = createClient(cookieStore);
   const data = await supabase
-    .from('document_v2')
+    .from("document_v2")
     .insert({
       id,
-      creator: session.user.name,
-      creator_email: session.user.email,
+      uid: session.user.id,
+      purview: 0, // 私有
       parent_id,
     })
     .select();
@@ -38,30 +37,32 @@ export async function createDoc(id: string, parent_id?: string) {
 }
 
 export async function delDoc(ids: string[]) {
+  const session = await auth();
+  if (!session?.user) return;
   const cookieStore = await cookies();
   const supabase = createClient(cookieStore);
-  const data = await supabase.from('document_v2').delete().in('id', ids);
-  return data;
-}
-
-export async function delTipDoc(id: string) {
-  const appId = process.env.NEXT_PUBLIC_TIPTAP_COLLAB_APP_ID as string;
-  const apiId = process.env.NEXT_PUBLIC_TIPTAP_API_ID as string;
-  const doc_name = `doc_${id}`;
-  const data = await fetch(
-    `https://${appId}.collab.tiptap.cloud/api/documents/${doc_name}`,
-    {
-      method: 'DELETE',
-      headers: {
-        Authorization: apiId,
-      },
-    }
-  );
+  const data = await supabase.from("document_v2").delete().in("id", ids);
   return data;
 }
 
 export async function updateTitle(id: string, title: string) {
+  const session = await auth();
+  if (!session?.user) return;
   const cookieStore = await cookies();
   const supabase = createClient(cookieStore);
-  return await supabase.from('document_v2').update({ title }).eq('id', id);
+  return await supabase.from("document_v2").update({ title }).eq("id", id);
+}
+
+export async function getDoc(id: string) {
+  const session = await auth();
+  if (!session?.user) return;
+  const cookieStore = await cookies();
+  const supabase = createClient(cookieStore);
+  const { data } = await supabase
+    .from("document_v2")
+    .select("*")
+    .eq("id", id)
+    .eq("uid", session.user.id)
+    .single();
+  return data as DocumentVO;
 }
